third_party/WebKit/LayoutTests/http/tests/inspector/network/font-face.html

<html>
<head>
<script src="/inspector/inspector-test.js"></script>
<script>
function createIFrame() {
    var iframe = document.createElement("iframe");
    iframe.src = "resources/font-face.html";
    document.body.appendChild(iframe);
}

function test() {
    function onRequest(event)
    {
        var request = event.data;
        if (request.name() === 'done') {
            InspectorTest.completeTest();
            return;
        }
        InspectorTest.addResult(event.type.toString() + ": " + request.name());
    }
    InspectorTest.networkManager.addEventListener(SDK.NetworkManager.Events.RequestStarted, onRequest);
    InspectorTest.networkManager.addEventListener(SDK.NetworkManager.Events.RequestFinished, onRequest);

    InspectorTest.evaluateInPage("createIFrame()");
}
</script>
</head>
<body onload="runTest()">
<p>Tests that a used font-face is reported and an unused font-face is not reported.</p>
</body>
</html>
