third_party/WebKit/LayoutTests/http/tests/inspector/network/font-face-expected.txt

Tests that a used font-face is reported and an unused font-face is not reported.


Symbol(RequestStarted): font-face.html
Symbol(RequestFinished): font-face.html
Symbol(RequestStarted): used.ttf
Symbol(RequestFinished): used.ttf
